refactor(gui): extract diagnostics chain into diagnostics.ts
